ENH: Upgrade VTK from v8.1.1 to v9.6.0 and modernize build

+++ Examples/antsVol.cxx
@@ -357,7 +357,7 @@ antsVolumetricRendering(itk::ants::CommandLineParser * parser)
   {
     vtkSmartPointer<vtkWindowToImageFilter> windowToImageFilter = vtkSmartPointer<vtkWindowToImageFilter>::New();
     windowToImageFilter->SetInput(renderWindow);
-    windowToImageFilter->SetScale(magnificationFactor);
+    windowToImageFilter->SetScale(static_cast<int>(magnificationFactor));
     windowToImageFilter->SetInputBufferTypeToRGBA();
     windowToImageFilter->Update();
